backend/pyproject.toml
[project]
name = "autofocus-backend"
version = "0.1.0"
description = "AutoFocus backend API for focus session management"
requires-python = ">=3.8.1"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "sqlalchemy==2.0.23",
    "python-multipart==0.0.6",
    "openai>=1.0.0",
    "httpx==0.27.2",
    "python-dotenv==1.0.0",
]

[tool.uv]
dev-dependencies = []